backend/requirements.txt
# Core
flask==2.3.2
werkzeug==2.3.6
Pillow==9.5.0

# Torch (compatible with Python 3.10)
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5

# DeOldify + fastai
fastai==1.0.61
scikit-image==0.19.3
matplotlib==3.5.3

# Face Restoration
gfpgan==1.3.8
basicsr==1.4.2
facexlib==0.2.5
realesrgan==0.3.0

# Misc
requests==2.31.0
tqdm==4.66.1
